add new screen for adhoc report
